@@ -1,6 +1,106 @@
+[build-system]
+requires = ["setuptools", "versioneer[toml]"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "pCrunch"
+dynamic = ["version"]
+description = "IO and Post Processing Interface for OpenFAST Results."
+readme = "README.rst"
+requires-python = ">=3.9"
+license = {text = "Apache-2.0"}
+keywords = ["wind", "turbine", "fatigue", "ultimate", "aeroelastic"]
+authors = [
+  {name = "NREL WISDEM Team", email = "systems.engineering@nrel.gov" }
+]
+maintainers = [
+    {name = "Jake Nunemaker", email ="jake.nunemaker@nrel.gov"},
+    {name = "Nikhar Abbas", email ="nikhar.abbas@nrel.gov"},
+]
+classifiers = [  # Optional
+  # How mature is this project? Common values are
+  #   3 - Alpha
+  #   4 - Beta
+  #   5 - Production/Stable
+  "Development Status :: 4 - Beta",
+
+  # Indicate who your project is intended for
+  "Intended Audience :: Science/Research",
+  "Topic :: Scientific/Engineering",
+
+  "License :: OSI Approved :: Apache Software License",
+
+  # Specify the Python versions you support here. In particular, ensure
+  # that you indicate you support Python 3. These classifiers are *not*
+  # checked by "pip install". See instead "python_requires" below.
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: C",
+  "Programming Language :: Fortran",
+]
+
+dependencies = ["matplotlib", "numpy", "scipy", "pyyaml", "seaborn", "fatpack"]
+
+# List additional groups of dependencies here (e.g. development
+# dependencies). Users will be able to install these using the "extras"
+# syntax, for example:
+#
+#   $ pip install sampleproject[dev]
+#
+# Similar to `dependencies` above, these must be valid existing
+# projects.
+[project.optional-dependencies] # Optional
+dev = ["pre-commit", "pylint", "flake8", "black", "isort"]
+test = ["pytest", "pytest-cov"]
+docs = ["sphinx", "sphinx_rtd_theme"]
+
+# List URLs that are relevant to your project
+#
+# This field corresponds to the "Project-URL" and "Home-Page" metadata fields:
+# https://packaging.python.org/specifications/core-metadata/#project-url-multiple-use
+# https://packaging.python.org/specifications/core-metadata/#home-page-optional
+#
+# Examples listed include a pattern for specifying where the package tracks
+# issues, where the source is hosted, where to say thanks to the package
+# maintainers, and where to support the project financially. The key is
+# what's used to render the link text on PyPI.
+[project.urls]  # Optional
+"Homepage" = "https://github.com/NREL/pCrunch"
+
+# This is configuration specific to the `setuptools` build backend.
+# If you are using a different build backend, you will need to change this.
+[tool.setuptools]
+include-package-data = true
+
+#[tool.setuptools.packages]
+#find = {}
+
+[tool.setuptools.packages.find]
+exclude = ["examples", "tests"]
+namespaces = true
+
+[tool.setuptools.package-data]
+# If there are data files included in your packages that need to be
+# installed, specify them here.
+"*" = ["*.out", "*.outb", "*.txt", "*.rst", "*.ipynb"]
+
+[tool.setuptools-git-versioning]
+enabled = true
+
+[tool.versioneer]
+VCS = "git"
+style = "pep440"
+versionfile_source = "pCrunch/_version.py"
+versionfile_build = "pCrunch/_version.py"
+tag_prefix = ""
+parentdir_prefix = "pCrunch-"
+
 [tool.black]
-line-length = 79
-target-version = ['py37']
+line-length = 88
+target-version = ['py311']
 include = '\.pyi?$'
 exclude = '''
 /(
